layer.conf: Add layers structure
Add a layers directory with subdirectories which will be included when
layers with names equal to them are included in BBPATH, which is usually
set through bblayers.conf.

Example:

meta-my-layer:
   |
   +-layers
       |
       +-foo
       |
       +-bar

If a layer foo is included in BBPATH then all .bb and .bbappend files
under layers/foo/ will be included. Otherwise said recipes will be
ignored.

Signed-off-by: Viktor Sjölind <vsjolind@luxoft.com>